requirements.txt
# Python dependencies for OCR extraction pipeline
# Install with: uv pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.28.0
pydantic>=2.0.0
instructor>=1.0.0
pandas>=2.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
tqdm>=4.65.0
einops>=0.7.0
qwen-vl-utils>=0.0.14

# Note: Also requires poppler-utils system package
# Ubuntu/Debian: sudo apt-get install poppler-utils
# macOS: brew install poppler
